Feat: 팔로우 버튼을 사용하는 곳에서 follow상태를 response하도록 수정

+++ src/main/java/com/finns/follow/controller/FollowController.java
@@ -62,8 +62,8 @@ public class FollowController {
 
     @GetMapping("/followingcheck")
     public ResponseEntity<Boolean> checkFollowing(
-            @RequestParam int user_no,
-            @RequestParam int to_user_no) {
+            @RequestParam long user_no,
+            @RequestParam long to_user_no) {
         boolean Following = followService.isFollowing(user_no, to_user_no);
         return ResponseEntity.ok(Following);
     }
